Cleanup billing route module
