+++ apps/labrinth/src/queue/mod.rs
@@ -1,5 +1,6 @@
 pub mod analytics;
 pub mod email;
+pub mod billing;
 pub mod maxmind;
 pub mod moderation;
 pub mod payouts;
