+++ apps/labrinth/src/routes/internal/medal.rs
@@ -9,8 +9,8 @@ use crate::database::models::users_redeemals::{
     Offer, RedeemalLookupFields, Status, UserRedeemal,
 };
 use crate::database::redis::RedisPool;
+use crate::queue::billing::try_process_user_redeemal;
 use crate::routes::ApiError;
-use crate::routes::internal::billing::try_process_user_redeemal;
 use crate::util::guards::medal_key_guard;
 
 pub fn config(cfg: &mut web::ServiceConfig) {
